requirements.txt
django~=4.2.11
psycopg2-binary
python-decouple
djangorestframework~=3.15.1
pillow
dj-rest-auth
drf-yasg
celery~=5.3.6
django-ckeditor~=6.7.1
django-filter~=24.2
djangorestframework-simplejwt
django-cors-headers
drf_social_oauth2
gunicorn
Flask~=3.0.2
stripe~=8.10.0
python-dotenv
djangochannelsrestframework
channels~=4.1.0
daphne
channels_redis
selenium #для тестов
asgiref~=3.8.1